refactor: separate implementation of angle and pi
